requirement.txt
opencv-python
openslide-python
pandas
pillow
matplotlib
numpy
ipython
tqdm
jupyterlab
ipywidgets
xxhash